Move internal/options/* -> internal/query/
